# Spatial Joins on Vectorized Geometries