--- a/Makefile
+++ b/Makefile
@@ -8,5 +8,5 @@ test: inplace
 	py.test geopandas
 
 clean:
-	rm -f geopandas/*.c geopandas/*.so
+	rm -f geopandas/vectorized.c geopandas/*.so geopandas/*.o
 	rm -rf build/ geopandas/__pycache__/ geopandas/*/__pycache__/