This adds a spatial join operation using the STRTree in GEOS.  All of the logic in implemented in C.  I had to add in the kvec.h library for dynamic arrays.

### Example

```python
In [1]: from geopandas.vectorized import VectorizedGeometry
   ...: from geopandas.vectorized import VectorizedGeometry, from_shapely, points_from_xy
   ...: import numpy as np
   ...: x = np.random.random(1000); y = np.random.random(1000)
   ...: import random
   ...: import shapely.geometry
   ...: triangles = [shapely.geometry.Polygon([(random.random(), random.random())
   ...:                                            for i in range(3)])
   ...:                  for _ in range(1000)]
   ...: T = from_shapely(triangles)
   ...: P = points_from_xy(x, y)
   ...: T
   ...: P
   ...: from geopandas.vectorized import VectorizedGeometry, from_shapely, points_from_xy, cysjoin
   ...: %time out = cysjoin(T.data, P.data, 'contains')
   ...: 
CPU times: user 3.32 s, sys: 12 ms, total: 3.33 s
Wall time: 3.34 s

In [2]: %load_ext ptime

In [3]: %ptime -n 4 out = cysjoin(T.data, P.data, 'contains')
Total serial time:   13.12 s
Total parallel time: 5.61 s
For a 2.34X speedup across 4 threads
```

I don't know yet if this is fast or slow.  It's an odd benchmark because most of the triangles overlap.  Also, should we be using preprared geometries for things like this?