@mrocklin 2017-07-30
This is also a good example of writing complex logic in C and only wrapping with Cython.