@mrocklin 2017-07-30
This appears to be slower than the current sjoin implementation