@mrocklin 2017-07-30
We're now about 10-20x faster than master for spatial joins on this benchmark.  I was using a smaller dataset before on the master comparison.  I've also switched to using prepared geometries.